--- src/checkQuotes.js
@@ -1,3 +1,0 @@
-const checkQuotes = (value) => value.indexOf('"') === 0 && value.lastIndexOf('"') === value.length - 1;
-
-export default checkQuotes;
