+++ src/formatFunctions.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
 /* eslint-disable import/prefer-default-export */
-import checkQuotes from './checkQuotes.js';
-import { cmcdHeader, keyTypes } from './constants.js';
-import { createError } from './error.js';
+import checkQuotes from './utils/checkQuotes.js';
+import { cmcdHeader, keyTypes } from './utils/constants.js';
+import { createError } from './utils/error.js';
 
 export const checkKeyInCorrectHeader = (header, key, errors) => {
   if (!cmcdHeader[header].includes(key)) {
@@ -15,6 +15,10 @@ export const isStringCorrect = (key, value, errors) => {
   if ((keyTypes[key] === 'string' && !checkQuotes(value)) || (keyTypes[key] === 'token' && checkQuotes(value))) {
     console.log('Incorrect format for ', key);
     errors.push(createError('incorrect-format'));
+    return true
+  }
+  else{
+    return false;
   }
 };
 
@@ -22,13 +26,18 @@ export const isBooleanCorrect = (key, value, errors) => {
   if ((keyTypes[key] === 'boolean') && value === 'true') {
     console.log('If the value is TRUE, the = and the value must be omitted');
     errors.push(createError('incorrect-format', key, value));
+    return false;
+  }
+  else{
+    return true;
   }
 };
 
-export const isSeaparetedCorrectly = (keyVal, errors) => {
+export const isSeparetedCorrectly = (keyVal, errors) => {
   if ((keyVal.split('=').length > 2) || ((keyVal.split('=').length === 1) && (keyTypes[keyVal] !== 'boolean'))) {
     errors.push(createError('incorrect-format'));
     console.log('key-value pair not separated by =.');
+    return false;
   }
   else {
     return true;
@@ -39,6 +48,10 @@ export const isKeyRepeated = (key, keys, errors) => {
   if (keys.includes(key)) {
     console.log(`The key '${key}' is repeated.`);
     errors.push(createError('duplicated-key', key));
+    return true;
+  }
+  else{
+    return false;
   }
 };
 
@@ -46,6 +59,10 @@ export const isHeaderRepeated = (header, headers, errors) => {
   if (headers.includes(header)) {
     console.log(`This header '${header}' is repeated.`);
     errors.push(createError('duplicated-header'));
+    return true
+  }
+  else{
+    return false;
   }
 };
 
@@ -56,7 +73,7 @@ export const noHeader= (headers, errors) => {
   }
 };
 
-export const emptyHeader = (keyVal, errors) => {
+export const isEmptyHeader = (keyVal, errors) => {
   if(keyVal === ''){
     console.log('Empty headers detected!');
     errors.push(createError('empty-header'));
